refactor(zones): split context resolution out of layoutregistry_assignments.cpp
