+++ libs/phosphor-zones/src/layoutregistry_contextresolve.cpp
@@ -1,0 +1,687 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: LGPL-2.1-or-later
+//
+// Per-context rule resolution for LayoutRegistry — the read side of the
+// assignment cascade (resolveAssignmentEntry, the gap / lock / default-
+// assignment / overlay / tiling-param resolvers, and the exact-context-rule
+// finders). Split from layoutregistry_assignments.cpp for file-size; the
+// assignment CRUD / mutators / query wrappers stay there.
+//
+// Phase 3b: per-context assignment is resolved on the unified Rule engine.
+// See layoutregistry_assignments.cpp for the resolution-model overview.
+
+#include <PhosphorZones/LayoutRegistry.h>
+
+#include "layoutregistry_rulehelpers_p.h"
+#include "zoneslogging.h"
+
+#include <PhosphorScreens/ScreenIdentity.h>
+#include <PhosphorScreens/VirtualScreen.h>
+
+#include <PhosphorRules/ContextRuleBridge.h>
+#include <PhosphorRules/MatchExpression.h>
+#include <PhosphorRules/RuleAction.h>
+#include <PhosphorRules/WindowQuery.h>
+#include <PhosphorRules/Rule.h>
+
+#include <QColor>
+
+#include <algorithm>
+#include <optional>
+
+namespace PhosphorZones {
+
+namespace PWR = PhosphorRules;
+
+// See layoutregistry_assignments.cpp for the RuleHelpers rationale — the
+// rule-shape classification / context helpers live in
+// layoutregistry_rulehelpers.cpp / _p.h and are shared across the registry TUs.
+using namespace RuleHelpers;
+
+namespace {
+
+/// True if @p match POSITIVELY pins @p field to a value — an `Equals` leaf on
+/// the field reachable without passing through a negation. Used to rank gap-rule
+/// specificity (a per-monitor ScreenId pin beats a per-mode Mode pin). Unlike
+/// `MatchExpression::referencesAnyField`, this deliberately does NOT count a leaf
+/// inside a `None{}` negation: a rule matching "every screen EXCEPT X" does not
+/// pin THIS screen, so it must not be ranked as a per-monitor override. Recurses
+/// through All/Any (a positive composite still constrains the field) but stops at
+/// None. Exotic non-`Equals` shapes fall through to the generic priority tier.
+bool matchPinsFieldPositively(const PWR::MatchExpression& match, PWR::Field field)
+{
+    switch (match.kind()) {
+    case PWR::MatchExpression::Kind::Leaf:
+        return match.predicate().field == field && match.predicate().op == PWR::Operator::Equals;
+    case PWR::MatchExpression::Kind::All:
+    case PWR::MatchExpression::Kind::Any:
+        for (const PWR::MatchExpression& child : match.children()) {
+            if (matchPinsFieldPositively(child, field)) {
+                return true;
+            }
+        }
+        return false;
+    case PWR::MatchExpression::Kind::None:
+        return false; // a negation does not positively pin the field
+    }
+    return false;
+}
+
+} // namespace
+
+// ── Rule-backed cascade resolution ──────────────────────────────────────────
+
+std::optional<AssignmentEntry> LayoutRegistry::resolveAssignmentEntry(const QString& screenId, int virtualDesktop,
+                                                                      const QString& activity) const
+{
+    // Resolution is per-slot. There are three independent slots — engine mode,
+    // snapping layout, tiling algorithm — each resolved separately, so a
+    // layout-only rule (a SetSnappingLayout or SetTilingAlgorithm with NO
+    // SetEngineMode) sets the layout for its engine in this context WITHOUT
+    // forcing the engine mode. The engine mode is the global default's (or
+    // another rule's); the layout slot is just filled.
+    //
+    // The winner per slot is simply the HIGHEST-PRIORITY matching rule carrying
+    // that slot's action, ties broken by list order. There is no specificity
+    // computation and no catch-all exclusion — `priority` is the only
+    // precedence value, and a user rule authored at a higher priority than a
+    // context assignment WILL win. This delegates to the shared
+    // RuleEvaluator::highestPriorityMatch — the one place the descending-
+    // priority, tie-break-by-list-order walk lives, so this resolver can never
+    // drift from the evaluator's own ordering.
+    //
+    // If no slot matched at all it's a genuine miss (nullopt) and the caller
+    // routes to the global default. If at least one slot matched, the entry's
+    // base is the engine-mode rule (mode + the layout tokens that rule itself
+    // carries, preserving mode-toggle losslessness); the per-slot layout
+    // winners then override the snapping-layout / tiling-algorithm fields. When
+    // only a layout rule matched (no engine-mode rule), the base is the global
+    // default for this context, so the resolved mode is the default's and the
+    // layout rule merely fills its slot.
+    //
+    // Hot-path cache via the shared revision-invalidated memoizer. The linear
+    // walk is O(N rules × M predicates per match); overlay/OSD callers issue it
+    // per cursor-move, and the connector + virtual-screen fallback chain in the
+    // public resolvers triples that on a miss. A @c nullopt is cached too — a
+    // genuine miss must not re-walk three times per cursor frame.
+    //
+    // Only the RULE-derived resolution is cached, keeping the cache a pure
+    // function of the rule set (its revision-invalidation contract). The global
+    // default — an external provider, not part of the rule set and not
+    // revision-tracked — is folded in AFTER the cache so a default-setting
+    // change (snapping toggled, a different default layout/algorithm) is
+    // reflected immediately, with no rule-set revision bump (a settings edit
+    // produces none). Caching the default-derived entry would otherwise pin a
+    // stale mode/layout on the layout-only path until the next rule edit. The
+    // fold-in is a handful of O(1) provider / per-context-slot reads, so the
+    // expensive priority walk stays memoized.
+    // Live tiled-window count for this context (nullopt when not actively
+    // tiling). It is NOT part of the rule set, so it cannot ride the cache's
+    // revision-invalidation contract like the global default does. Instead it
+    // is folded into the cache KEY (the `mode` slot of this cache, which for the
+    // assignment resolver carries only the count + orientation composite below) so a
+    // count change yields a fresh entry rather than a stale hit, while
+    // count-independent callers (overlay / OSD per cursor-move, where the count is
+    // steady) keep hitting the cache.
+    const std::optional<int> tiledCount =
+        m_tiledWindowCountProvider ? m_tiledWindowCountProvider(screenId, virtualDesktop, activity) : std::nullopt;
+    // Both the tiled count and the screen orientation are non-rule-set inputs the
+    // stamped query depends on, so both must ride the cache KEY (the assignment
+    // resolver stamps orientation but not activeLayout, so no "al:" component here).
+    const QString orientationToken = screenOrientationToken(screenId);
+    const QString countCacheKey = (tiledCount ? (QLatin1String("twc:") + QString::number(*tiledCount)) : QString())
+        + QLatin1String("|or:") + orientationToken;
+
+    const std::optional<RuleSlotResolution> rules = resolveCachedContext(
+        m_contextResolveCache, m_contextResolveCacheRevision, screenId, virtualDesktop, activity, countCacheKey,
+        [&]() -> std::optional<RuleSlotResolution> {
+            PWR::WindowQuery query = makeContextQuery(screenId, virtualDesktop, activity);
+            // The tiled count is supplied ONLY to the assignment resolver's
+            // query, because algorithm switching is its only intended use. The
+            // gap / lock / overlay / default-assignment resolvers build their own
+            // makeContextQuery() without it, so a Field::TiledWindowCount predicate
+            // on a rule carrying one of those actions stays inert (the field is
+            // absent there) by design.
+            query.tiledWindowCount = tiledCount;
+            // Orientation is layout-independent, so it is stamped on EVERY context
+            // query (unlike tiledWindowCount / activeLayout) — an orientation rule
+            // can drive the assignment itself (portrait monitor → different layout).
+            // Reuse the token already captured for the cache key (mirrors the
+            // gap/lock/overlay lambdas) rather than re-reading the provider.
+            query.screenOrientation = orientationToken;
+            // Field::ActiveLayout is deliberately NOT stamped here: the active
+            // layout IS this resolver's output, and reading it (assignmentIdForScreen
+            // → resolveAssignmentEntry) would recurse. So an ActiveLayout rule cannot
+            // drive the layout assignment (a circular request); it is populated only
+            // by the post-assignment resolvers (gap / lock / overlay).
+
+            // Highest-priority matching rule per slot wins (ties by list order).
+            // A rule that REFERENCES Field::ActiveLayout is structurally excluded from
+            // the assignment path: activeLayout is left unstamped above, so such a
+            // rule evaluates against a placeholder empty value. A positive leaf
+            // (ActiveLayout Equals X) never matches that placeholder, but a negated
+            // predicate (None{ActiveLayout Equals X}) WOULD spuriously match and force
+            // a wrong assignment. Dropping any ActiveLayout-referencing rule here keeps
+            // ActiveLayout a strictly post-assignment field regardless of predicate
+            // polarity, not merely by relying on the empty-value coincidence.
+            const auto slotMatch = [&](bool (*carriesSlot)(const PWR::Rule&)) -> const PWR::Rule* {
+                return m_evaluator->highestPriorityMatch(query, [carriesSlot](const PWR::Rule& rule) {
+                    return carriesSlot(rule) && !rule.match.referencesAnyField({PWR::Field::ActiveLayout});
+                });
+            };
+
+            const PWR::Rule* modeRule = slotMatch(hasEngineModeAction);
+            const PWR::Rule* snapRule = slotMatch(hasSnappingLayoutAction);
+            const PWR::Rule* algoRule = slotMatch(hasTilingAlgorithmAction);
+
+            if (modeRule == nullptr && snapRule == nullptr && algoRule == nullptr) {
+                return std::nullopt; // genuine miss — the caller routes to the default
+            }
+
+            // The engine-mode rule decides the mode (and carries its own layout
+            // tokens, preserving mode-toggle losslessness); the per-slot layout
+            // winners fill their own field. Each slot independently took the
+            // highest-priority matching rule, so the layout slots track their
+            // own winner regardless of the engine-mode rule.
+            RuleSlotResolution resolved;
+            if (modeRule != nullptr) {
+                resolved.modeEntry = entryFromRuleMatchActions(*modeRule);
+            }
+            if (snapRule != nullptr) {
+                resolved.snappingLayout = entryFromRuleMatchActions(*snapRule).snappingLayout;
+            }
+            if (algoRule != nullptr) {
+                resolved.tilingAlgorithm = entryFromRuleMatchActions(*algoRule).tilingAlgorithm;
+            }
+            return resolved;
+        });
+
+    if (!rules) {
+        return std::nullopt; // genuine miss — the caller routes to the default
+    }
+
+    // Base: the engine-mode rule's decoded entry. With no engine-mode rule, the
+    // live default for this context supplies the mode — a layout-only rule must
+    // not force one. Resolved here, OUTSIDE the cache, so it tracks settings.
+    AssignmentEntry entry = rules->modeEntry
+        ? *rules->modeEntry
+        : resolveDefaultAssignmentEntryForContext(screenId, virtualDesktop, activity);
+
+    // Per-slot layout winners override their field.
+    if (rules->snappingLayout) {
+        entry.snappingLayout = *rules->snappingLayout;
+    }
+    if (rules->tilingAlgorithm) {
+        entry.tilingAlgorithm = *rules->tilingAlgorithm;
+    }
+    return entry;
+}
+
+ContextGapOverride LayoutRegistry::resolveContextGaps(const QString& screenId, int virtualDesktop,
+                                                      const QString& activity, const QString& mode) const
+{
+    // This returns CONTEXT OVERRIDE rules only — the per-screen, per-desktop and
+    // per-activity gap rules that sit ABOVE the per-layout tier in the geometry
+    // cascade (getEffectiveInnerGap / getEffectiveOuterGaps tier 1). It is NOT
+    // the cascade's default tier: the global default gap lives on the managed
+    // catch-all baseline rule and is read separately, BY ID, through the
+    // consumer's global inner/outer gap settings at the cascade's default tier
+    // (tier 3), with the per-layout override sitting between this override layer
+    // and that default. The clean tiering is: context overrides (here) →
+    // per-layout → global default (baseline by id) → compile default.
+    //
+    // Unlike resolveAssignmentEntry (single winning engine-mode rule), gap
+    // overrides are read PER SLOT from the evaluator's ResolvedActions, so a
+    // gap-only rule and, say, a separate engine-mode rule on the same context
+    // compose — and two gap rules touching different slots both apply. No
+    // engine-mode gate and no isValid() filter: a gap-only context rule is a
+    // first-class override here.
+    if (!m_evaluator) {
+        return ContextGapOverride{};
+    }
+
+    // The active layout AND the screen orientation are folded into the cache key
+    // (see contextCacheKeyToken) so a Field::ActiveLayout / Field::ScreenOrientation
+    // gap rule refreshes when either changes; both are stamped onto the query below.
+    // Safe from recursion: assignmentIdForScreen routes through resolveAssignmentEntry,
+    // which never calls back into the gap resolver.
+    const QString activeLayoutId = assignmentIdForScreen(screenId, virtualDesktop, activity);
+    const QString orientationToken = screenOrientationToken(screenId);
+
+    // Hot-path cache via the shared revision-invalidated memoizer: the geometry
+    // path resolves the same context twice per op (zone padding + outer gaps)
+    // and N× inside a multi-zone snap, all with identical arguments.
+    return resolveCachedContext(
+        m_contextGapCache, m_contextGapCacheRevision, screenId, virtualDesktop, activity,
+        contextCacheKeyToken(mode, activeLayoutId, orientationToken), [&]() -> ContextGapOverride {
+            ContextGapOverride gaps;
+            // Thread the placement mode into the query so a per-mode `Mode
+            // Equals "snapping"/"tiling"` gap rule resolves for the asking
+            // engine and stays inert for the other.
+            PWR::WindowQuery query = makeContextQuery(screenId, virtualDesktop, activity, mode);
+            query.screenOrientation = orientationToken;
+            query.activeLayout = activeLayoutId;
+
+            // Resolve each gap slot from the highest-priority matching rule that
+            // carries that slot's action. The CATCH-ALL managed baseline rule is
+            // EXCLUDED because it is the cascade's DEFAULT TIER, not a context
+            // override: it carries the GLOBAL default gap values and is read by id
+            // through the consumer's global inner/outer gap settings at the
+            // geometry cascade's default tier, with the per-layout override sitting
+            // between this override layer and that default. Were the baseline
+            // included here, its
+            // catch-all match would fill every gap slot with the global default and
+            // masquerade as a top-tier context override, shadowing the per-layout
+            // tier that must sit below context overrides. SCREEN-scoped gap rules
+            // (per-monitor overrides authored via the Appearance page's monitor
+            // scope) have a SPECIFIC match, so they are NOT catch-all and DO
+            // participate here as context overrides. Per slot, the winner is chosen
+            // by MATCH SPECIFICITY first (ScreenId-pinned > Mode-pinned > other),
+            // with priority breaking ties within a tier — so a per-monitor override
+            // outranks a global per-mode gap regardless of their raw priorities (see
+            // winningAction below). The catch-all baseline is excluded by the
+            // per-slot "carries this slot, not the catch-all baseline" filter.
+            const PWR::ActionRegistry& registry = PWR::ActionRegistry::instance();
+            const auto winningAction = [this, &query,
+                                        &registry](QLatin1StringView slot) -> std::optional<PWR::RuleAction> {
+                const QString slotId = QString(slot);
+                const auto carries = [&registry, &slotId](const PWR::Rule& rule) {
+                    if (rule.managed && rule.match.isCatchAll()) {
+                        return false;
+                    }
+                    for (const PWR::RuleAction& a : rule.actions) {
+                        if (registry.slotFor(a) == slotId) {
+                            return true;
+                        }
+                    }
+                    return false;
+                };
+                // Resolve the slot by MATCH SPECIFICITY, not raw priority: a
+                // per-monitor (ScreenId-pinned) gap override is more specific than
+                // a per-mode (Mode-pinned) one, which is more specific than any
+                // other context rule carrying the slot. Try each tier in turn so a
+                // per-monitor override beats a global per-mode gap (the v4 cascade,
+                // where the monitor-specific value won), with priority breaking ties
+                // within a tier. Specificity uses matchPinsFieldPositively (an Equals
+                // leaf not under a negation), so a "every screen except X" negated
+                // rule is NOT mis-ranked as a per-monitor override and instead falls
+                // through to the generic priority tier. Only the gap cascade is
+                // specificity-ordered; appearance slots resolve by priority alone in
+                // the effect.
+                const auto carriesScreenPinned = [&carries](const PWR::Rule& rule) {
+                    return carries(rule) && matchPinsFieldPositively(rule.match, PWR::Field::ScreenId);
+                };
+                const auto carriesModePinned = [&carries](const PWR::Rule& rule) {
+                    return carries(rule) && matchPinsFieldPositively(rule.match, PWR::Field::Mode);
+                };
+                const PWR::Rule* rule = m_evaluator->highestPriorityMatch(query, carriesScreenPinned);
+                if (rule == nullptr) {
+                    rule = m_evaluator->highestPriorityMatch(query, carriesModePinned);
+                }
+                if (rule == nullptr) {
+                    rule = m_evaluator->highestPriorityMatch(query, carries);
+                }
+                if (rule == nullptr) {
+                    return std::nullopt;
+                }
+                for (const PWR::RuleAction& a : rule->actions) {
+                    if (registry.slotFor(a) == slotId) {
+                        return a;
+                    }
+                }
+                return std::nullopt;
+            };
+            const auto readInt = [&winningAction](QLatin1StringView slot, std::optional<int>& out) {
+                if (const auto action = winningAction(slot)) {
+                    out = action->params.value(PWR::ActionParam::Value).toInt();
+                }
+            };
+            const auto readBool = [&winningAction](QLatin1StringView slot, std::optional<bool>& out) {
+                if (const auto action = winningAction(slot)) {
+                    out = action->params.value(PWR::ActionParam::Value).toBool();
+                }
+            };
+            readInt(PWR::ActionSlot::InnerGap, gaps.innerGap);
+            readInt(PWR::ActionSlot::OuterGap, gaps.outerGap);
+            readBool(PWR::ActionSlot::UsePerSideOuterGap, gaps.usePerSideOuterGap);
+            readInt(PWR::ActionSlot::OuterGapTop, gaps.outerGapTop);
+            readInt(PWR::ActionSlot::OuterGapBottom, gaps.outerGapBottom);
+            readInt(PWR::ActionSlot::OuterGapLeft, gaps.outerGapLeft);
+            readInt(PWR::ActionSlot::OuterGapRight, gaps.outerGapRight);
+            return gaps;
+        });
+}
+
+bool LayoutRegistry::resolveContextLocked(const QString& screenId, int virtualDesktop, const QString& activity) const
+{
+    // Mirror resolveContextGaps: a per-slot read off the evaluator's
+    // ResolvedActions, not the single-winner assignment walk. The Locked slot
+    // is filled by the highest-priority matching context rule carrying a
+    // LockContext action; we report its boolean value. No engine-mode gate —
+    // a lock-only context rule is a first-class overlay.
+    if (!m_evaluator) {
+        return false;
+    }
+
+    // Active layout + orientation folded into the cache key + stamped onto the
+    // query, so a Field::ActiveLayout / ScreenOrientation lock rule works and
+    // refreshes when either changes.
+    const QString activeLayoutId = assignmentIdForScreen(screenId, virtualDesktop, activity);
+    const QString orientationToken = screenOrientationToken(screenId);
+
+    // Hot-path cache via the shared revision-invalidated memoizer: the lock
+    // check runs per cursor-move while a selector is open and on every
+    // layout-switch attempt.
+    return resolveCachedContext(m_contextLockCache, m_contextLockCacheRevision, screenId, virtualDesktop, activity,
+                                contextCacheKeyToken(QString(), activeLayoutId, orientationToken), [&]() -> bool {
+                                    PWR::WindowQuery query = makeContextQuery(screenId, virtualDesktop, activity);
+                                    query.screenOrientation = orientationToken;
+                                    query.activeLayout = activeLayoutId;
+                                    const PWR::ResolvedActions resolved = m_evaluator->resolve(query);
+                                    if (const auto action = resolved.slot(QString(PWR::ActionSlot::Locked))) {
+                                        return action->params.value(PWR::ActionParam::Value).toBool();
+                                    }
+                                    return false;
+                                });
+}
+
+std::optional<bool> LayoutRegistry::resolveContextDefaultAssignment(const QString& screenId, int virtualDesktop,
+                                                                    const QString& activity) const
+{
+    // Mirror resolveContextLocked: a per-slot read off the evaluator's
+    // ResolvedActions, not the single-winner assignment walk. The
+    // DefaultAssignment slot is filled by the highest-priority matching context
+    // rule carrying a DefaultLayoutAssignment action; we report its boolean
+    // value (true = allow / force the default through, false = suppress). No
+    // engine-mode gate — a default-assignment-only context rule is a first-class
+    // overlay. std::nullopt means "no override rule" — the caller follows the
+    // global setting.
+    if (!m_evaluator) {
+        return std::nullopt;
+    }
+
+    // Orientation folded into the cache key so a Field::ScreenOrientation
+    // default-assignment rule refreshes on rotation (activeLayout is deliberately
+    // NOT stamped/folded here — see the recursion note in the lambda below).
+    const QString orientationToken = screenOrientationToken(screenId);
+    return resolveCachedContext(
+        m_contextDefaultAssignmentCache, m_contextDefaultAssignmentCacheRevision, screenId, virtualDesktop, activity,
+        contextCacheKeyToken(QString(), QString(), orientationToken), [&]() -> std::optional<bool> {
+            PWR::WindowQuery query = makeContextQuery(screenId, virtualDesktop, activity);
+            query.screenOrientation = orientationToken; // reuse the cache-key token
+            // Field::ActiveLayout NOT stamped here: this resolver is part of the
+            // assignment cascade (assignmentIdForScreen reaches it via
+            // resolveDefaultAssignmentEntryForContext), so stamping the active
+            // layout here would recurse. See resolveAssignmentEntry.
+            //
+            // Because activeLayout is unstamped, an ActiveLayout-referencing rule
+            // must be structurally excluded from this no-stamp resolver too (the
+            // exact symmetry resolveAssignmentEntry's slotMatch enforces): a negated
+            // None{ActiveLayout Equals X} predicate would otherwise spuriously match
+            // the empty placeholder and wrongly force/suppress the default. Use a
+            // filtered highestPriorityMatch rather than the unfiltered resolve().
+            const PWR::Rule* rule = m_evaluator->highestPriorityMatch(query, [](const PWR::Rule& r) {
+                if (r.match.referencesAnyField({PWR::Field::ActiveLayout})) {
+                    return false;
+                }
+                for (const PWR::RuleAction& action : r.actions) {
+                    if (action.type == QLatin1String(PWR::ActionType::DefaultLayoutAssignment)) {
+                        return true;
+                    }
+                }
+                return false;
+            });
+            if (rule != nullptr) {
+                for (const PWR::RuleAction& action : rule->actions) {
+                    if (action.type == QLatin1String(PWR::ActionType::DefaultLayoutAssignment)) {
+                        return action.params.value(PWR::ActionParam::Value).toBool();
+                    }
+                }
+            }
+            return std::nullopt;
+        });
+}
+
+AssignmentEntry LayoutRegistry::resolveDefaultAssignmentEntryForContext(const QString& screenId, int virtualDesktop,
+                                                                        const QString& activity) const
+{
+    // Single cascade-miss tail for every per-context resolver. The per-context
+    // DefaultLayoutAssignment override (if present) wins over the global
+    // suppress baseline:
+    //   - false (suppress) → invalid entry: this context gets no default.
+    //   - true  (allow)    → raw synth: force the default through even when the
+    //                        global suppress setting is on.
+    //   - no override      → the global-gated resolveDefaultAssignmentEntry.
+    if (const auto contextOverride = resolveContextDefaultAssignment(screenId, virtualDesktop, activity)) {
+        return *contextOverride ? resolveDefaultAssignmentEntryRaw() : AssignmentEntry{};
+    }
+    return resolveDefaultAssignmentEntry();
+}
+
+ContextOverlayOverride LayoutRegistry::resolveContextOverlay(const QString& screenId, int virtualDesktop,
+                                                             const QString& activity) const
+{
+    // Per-slot read across all matching context rules (mirrors resolveContextGaps):
+    // independent overlay-shader / overlay-style rules compose, and
+    // each populated field overrides the active layout's own value at the overlay
+    // build site. No engine-mode gate — an overlay-only context rule is first-class.
+    if (!m_evaluator) {
+        return ContextOverlayOverride{};
+    }
+
+    // Active layout + orientation folded into the cache key + stamped onto the
+    // query, so a Field::ActiveLayout / ScreenOrientation overlay rule works and
+    // refreshes when either changes.
+    const QString activeLayoutId = assignmentIdForScreen(screenId, virtualDesktop, activity);
+    const QString orientationToken = screenOrientationToken(screenId);
+
+    return resolveCachedContext(
+        m_contextOverlayCache, m_contextOverlayCacheRevision, screenId, virtualDesktop, activity,
+        contextCacheKeyToken(QString(), activeLayoutId, orientationToken), [&]() -> ContextOverlayOverride {
+            ContextOverlayOverride overlay;
+            PWR::WindowQuery query = makeContextQuery(screenId, virtualDesktop, activity);
+            query.screenOrientation = orientationToken;
+            query.activeLayout = activeLayoutId;
+            const PWR::ResolvedActions resolved = m_evaluator->resolve(query);
+
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayShader))) {
+                const QString id = action->params.value(PWR::ActionParam::EffectId).toString();
+                if (!id.isEmpty()) {
+                    overlay.shaderId = id;
+                    // Optional shader uniform overrides — empty when the rule
+                    // overrides only the shader id (shader defaults apply).
+                    overlay.shaderParams = action->params.value(PWR::ActionParam::Params).toObject().toVariantMap();
+                }
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayStyle))) {
+                // Wire token → OverlayDisplayMode int so consumers compare against
+                // the same enum Layout::overlayDisplayMode() exposes (0 =
+                // ZoneRectangles, 1 = LayoutPreview).
+                const QString token = action->params.value(PWR::ActionParam::Value).toString();
+                if (token == PWR::OverlayStyleToken::Rectangles) {
+                    overlay.style = 0;
+                } else if (token == PWR::OverlayStyleToken::Preview) {
+                    overlay.style = 1;
+                }
+            }
+            // Appearance overrides — each fills its own optional so an unset
+            // property falls through to the global config value at the consumer.
+            // Colours are parsed from the `#AARRGGBB` wire hex; QColor reads a
+            // 9-digit hex alpha-first, matching the picker's toHexArgb output. The
+            // descriptor validators reject malformed values at load; the guards here
+            // are defense in depth so a hand-edited rules.json that slipped a bad value
+            // can only fall through to config or a sane bound, never apply a broken
+            // override: colours are validity-checked, opacities clamped to [0, 1], and
+            // border width / radius floored at 0 (negatives are nonsensical; the load
+            // validator enforces the upper bounds, which the shared constants own).
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayHighlightColor))) {
+                if (const QColor c(action->params.value(PWR::ActionParam::Value).toString()); c.isValid()) {
+                    overlay.highlightColor = c;
+                }
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayInactiveColor))) {
+                if (const QColor c(action->params.value(PWR::ActionParam::Value).toString()); c.isValid()) {
+                    overlay.inactiveColor = c;
+                }
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayBorderColor))) {
+                if (const QColor c(action->params.value(PWR::ActionParam::Value).toString()); c.isValid()) {
+                    overlay.borderColor = c;
+                }
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayActiveOpacity))) {
+                overlay.activeOpacity = qBound(0.0, action->params.value(PWR::ActionParam::Value).toDouble(), 1.0);
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayInactiveOpacity))) {
+                overlay.inactiveOpacity = qBound(0.0, action->params.value(PWR::ActionParam::Value).toDouble(), 1.0);
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayBorderWidth))) {
+                overlay.borderWidth = std::max(0, action->params.value(PWR::ActionParam::Value).toInt());
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayBorderRadius))) {
+                overlay.borderRadius = std::max(0, action->params.value(PWR::ActionParam::Value).toInt());
+            }
+            if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverlayShowZoneNumbers))) {
+                overlay.showZoneNumbers = action->params.value(PWR::ActionParam::Value).toBool();
+            }
+            return overlay;
+        });
+}
+
+ContextTilingParams LayoutRegistry::resolveContextTilingParams(const QString& screenId, int virtualDesktop,
+                                                               const QString& activity) const
+{
+    if (!m_evaluator) {
+        return {};
+    }
+    // Per-slot read (mirrors resolveContextGaps), but NOT cached: this runs on
+    // screen / layout changes via the daemon's updateAutotileScreens, not the hot
+    // per-cursor path. Being uncached lets us stamp the active layout AND the
+    // screen orientation onto the query without folding either into a cache key
+    // (no cached entry to go stale). Safe from recursion: assignmentIdForScreen
+    // routes through resolveAssignmentEntry, which never calls this resolver.
+    PWR::WindowQuery query = makeContextQuery(screenId, virtualDesktop, activity);
+    stampScreenOrientation(query, screenId);
+    query.activeLayout = assignmentIdForScreen(screenId, virtualDesktop, activity);
+    const PWR::ResolvedActions resolved = m_evaluator->resolve(query);
+
+    ContextTilingParams params;
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::MaxWindows))) {
+        params.maxWindows = action->params.value(PWR::ActionParam::Value).toInt();
+    }
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::SplitRatio))) {
+        params.splitRatio = action->params.value(PWR::ActionParam::Value).toDouble();
+    }
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::MasterCount))) {
+        params.masterCount = action->params.value(PWR::ActionParam::Value).toInt();
+    }
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::InsertPosition))) {
+        // Wire token → AutotileInsertPosition int (End 0 / AfterFocused 1 / AsMaster 2),
+        // the same value the per-screen config store holds.
+        const QString token = action->params.value(PWR::ActionParam::Value).toString();
+        if (token == PWR::InsertPositionToken::End) {
+            params.insertPosition = 0;
+        } else if (token == PWR::InsertPositionToken::AfterFocused) {
+            params.insertPosition = 1;
+        } else if (token == PWR::InsertPositionToken::AsMaster) {
+            params.insertPosition = 2;
+        }
+    }
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::OverflowBehavior))) {
+        // Wire token → AutotileOverflowBehavior int (Float 0 / Unlimited 1).
+        const QString token = action->params.value(PWR::ActionParam::Value).toString();
+        if (token == PWR::OverflowBehaviorToken::Float) {
+            params.overflowBehavior = 0;
+        } else if (token == PWR::OverflowBehaviorToken::Unlimited) {
+            params.overflowBehavior = 1;
+        }
+    }
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::DragBehavior))) {
+        // Wire token → AutotileDragBehavior int (Float 0 / Reorder 1).
+        const QString token = action->params.value(PWR::ActionParam::Value).toString();
+        if (token == PWR::DragBehaviorToken::Float) {
+            params.dragBehavior = 0;
+        } else if (token == PWR::DragBehaviorToken::Reorder) {
+            params.dragBehavior = 1;
+        }
+    }
+    if (const auto action = resolved.slot(QString(PWR::ActionSlot::AlgorithmParams))) {
+        // Target algorithm token + free-form custom-param blob (mirrors the
+        // overlay shader-uniform override). The daemon applies the params only
+        // when the target matches the screen's effective algorithm.
+        params.algorithmParamTarget = action->params.value(PWR::ActionParam::Algorithm).toString();
+        params.algorithmParams = action->params.value(PWR::ActionParam::Params).toObject().toVariantMap();
+    }
+    return params;
+}
+
+bool LayoutRegistry::hasExactContextRule(const QString& screenId, int virtualDesktop, const QString& activity) const
+{
+    return findExactContextRule(screenId, virtualDesktop, activity) != nullptr;
+}
+const PhosphorRules::Rule* LayoutRegistry::findExactContextRule(const QString& screenId, int virtualDesktop,
+                                                                const QString& activity) const
+{
+    // The deterministic v5 derivation lets us look up a stored assignment by
+    // its identity tuple — the bridge guarantees identical tuples produce
+    // identical ids. We scan the rule list once for the id (the linear walk
+    // is unavoidable here because callers need a stable in-set pointer, not
+    // the value-copy `ruleById` returns), then guard with
+    // hasEngineModeAction + matchIsExactContext so a hand-edited match that
+    // no longer satisfies the canonical context shape can never be returned
+    // even if its id happens to match the deterministic derivation.
+    //
+    // The win over the previous implementation is the predicate cost: the
+    // old scan called matchIsExactContext on EVERY rule's match expression;
+    // this scan compares ids first and only evaluates the context-shape
+    // predicate on the unique candidate.
+    //
+    // O(N) intentional: a hash lookup would need RuleSet to expose a
+    // pointer-returning accessor into its in-set storage with a documented
+    // dangle-on-setRules contract — too sharp an edge for the win at the
+    // rule counts we care about (<= ~1000 rules per profile). Revisit only
+    // if rule counts grow well into the thousands.
+    // Two-pass scan: prefer the deterministic-id rule (cheap id compare,
+    // the canonical shape the bridge produces). If no rule has the
+    // canonical id, fall back to a shape-based scan that picks up
+    // user-authored PURE-assignment rules whose UUIDs were generated by
+    // the settings UI (ruletemplates.cpp's `QUuid::createUuid()`
+    // path). The shape fallback gates on `isPureAssignmentRule` — a
+    // user rule that ALSO carries SetOpacity / OverrideAnimation* /
+    // Float / Exclude / LockContext alongside the assignment slots is
+    // intentionally NOT claimed here. Admitting it would feed it through
+    // upsertAssignmentRule / assignLayout / applyBatchAssignments which
+    // rebuild via makeAssignmentActions and emit only the three slot
+    // actions — silently stripping the user's other actions. Falling
+    // through to the addRule branch instead preserves the user's data
+    // (the duplicate-shadow at the same cascade band is a known limit;
+    // documented at upsertAssignmentRule's contract).
+    const QUuid candidateId = PWR::ContextRuleBridge::assignmentRuleIdFor(screenId, virtualDesktop, activity);
+    const PWR::Rule* shapeMatch = nullptr;
+    for (const PWR::Rule& rule : m_ruleStore->ruleSet().rules()) {
+        if (rule.id == candidateId) {
+            if (!hasEngineModeAction(rule) || !matchIsExactContext(rule.match, screenId, virtualDesktop, activity)) {
+                // Deterministic-id rule exists but its match shape was
+                // hand-edited away from the canonical form. Return
+                // nothing so callers don't act on a malformed rule.
+                return nullptr;
+            }
+            return &rule;
+        }
+        // Remember the first PURE-assignment rule we see — we only
+        // return it if no deterministic-id rule exists in the set.
+        if (shapeMatch == nullptr && isPureAssignmentRule(rule)
+            && matchIsExactContext(rule.match, screenId, virtualDesktop, activity)) {
+            shapeMatch = &rule;
+        }
+    }
+    return shapeMatch;
+}
+
+QUuid LayoutRegistry::exactContextRuleId(const QString& screenId, int virtualDesktop, const QString& activity) const
+{
+    const PWR::Rule* rule = findExactContextRule(screenId, virtualDesktop, activity);
+    return rule ? rule->id : QUuid();
+}
+
+} // namespace PhosphorZones
